重写了部分战斗系统的代码

+++ setup.py
@@ -1,21 +1,20 @@
 from distutils.core import setup
 from Cython.Build import cythonize
-import glob
+from glob import glob
 import os
 import shutil
 
+#py setup.py build_ext --inplace
 #pyinstaller -i icon.ico main.py
 #pip list --outdated --format=freeze | grep -v '^\-e' | cut -d = -f 1  | xargs -n1 pip install -U
 
-#是否编译Source中的游戏本体
-compile_py_files_in_Source = False
 debug_c = False
 
 #生成游戏本体源代码的c和pyd文件
-if compile_py_files_in_Source:
-    for path in glob.glob(r'Source/*.py'):
-        setup(ext_modules=cythonize(path))
-        if not debug_c:
-            os.remove(path.replace(".pyx",".c"))
+for path in glob(r'Source/*.py'):
+    setup(ext_modules=cythonize(path))
+    #删除.c文件
+    if not debug_c: os.remove(path.replace(".pyx",".c"))
+
 #删除build文件夹
-shutil.rmtree('build')
+if os.path.exists('build'): shutil.rmtree('build')
